requirements.txt
numpy
pyqt5
pillow
zwoasi
opencv-python
pyzmq